requirements/requirements.txt
python-telegram-bot[job-queue]==22.3
SQLAlchemy==2.0.43
asyncpg==0.30.0
httpx==0.28.1
python-dotenv==1.1.1
greenlet==3.2.4